Replace isWhiteSpaceLike, isWhiteSpaceSingleLine and isLineBreak
- Modularize Rust source code
- Create a types module
