+++ TypeScript-replace/src/compiler/scanner.ts
@@ -399,44 +399,17 @@ namespace ts {
     }
 
     export function isWhiteSpaceLike(ch: number): boolean {
-        return isWhiteSpaceSingleLine(ch) || isLineBreak(ch);
+        return RustyTypeScript.is_white_space_like(ch);
     }
 
     /** Does not include line breaks. For that, see isWhiteSpaceLike. */
     export function isWhiteSpaceSingleLine(ch: number): boolean {
-        // Note: nextLine is in the Zs space, and should be considered to be a whitespace.
-        // It is explicitly not a line-break as it isn't in the exact set specified by EcmaScript.
-        return ch === CharacterCodes.space ||
-            ch === CharacterCodes.tab ||
-            ch === CharacterCodes.verticalTab ||
-            ch === CharacterCodes.formFeed ||
-            ch === CharacterCodes.nonBreakingSpace ||
-            ch === CharacterCodes.nextLine ||
-            ch === CharacterCodes.ogham ||
-            ch >= CharacterCodes.enQuad && ch <= CharacterCodes.zeroWidthSpace ||
-            ch === CharacterCodes.narrowNoBreakSpace ||
-            ch === CharacterCodes.mathematicalSpace ||
-            ch === CharacterCodes.ideographicSpace ||
-            ch === CharacterCodes.byteOrderMark;
-      }
+        return RustyTypeScript.is_white_space_single_line(ch);
+    }
 
-      export function isLineBreak(ch: number): boolean {
-          // ES5 7.3:
-          // The ECMAScript line terminator characters are listed in Table 3.
-          //     Table 3: Line Terminator Characters
-          //     Code Unit Value     Name                    Formal Name
-          //     \u000A              Line Feed               <LF>
-          //     \u000D              Carriage Return         <CR>
-          //     \u2028              Line separator          <LS>
-          //     \u2029              Paragraph separator     <PS>
-          // Only the characters in Table 3 are treated as line terminators. Other new line or line
-          // breaking characters are treated as white space but not as line terminators.
-
-          return ch === CharacterCodes.lineFeed ||
-              ch === CharacterCodes.carriageReturn ||
-              ch === CharacterCodes.lineSeparator ||
-              ch === CharacterCodes.paragraphSeparator;
-      }
+    export function isLineBreak(ch: number): boolean {
+        return RustyTypeScript.is_line_break(ch);
+    }
 
       function isDigit(ch: number): boolean {
           return ch >= CharacterCodes._0 && ch <= CharacterCodes._9;
